requirements.txt
fastapi
uvicorn
pydantic
numpy
pandas
scikit-learn
faiss-cpu
tiktoken
ragas
datasets
nltk
pyyaml
rich
loguru
typer
httpx
orjson
openai
tenacity
black
ruff
mypy
pytest
pytest-cov
beautifulsoup4
lxml
requests
